[FRONT] Add new model Profile linked to User
