+++ mon-pix/app/models/user.js
@@ -18,6 +18,7 @@ export default class User extends Model {
 
   // includes
   @belongsTo('is-certifiable') isCertifiable;
+  @belongsTo('profile', { async: false }) profile;
   @belongsTo('pix-score') pixScore;
   @hasMany('campaign-participation') campaignParticipations;
   @hasMany('certification') certifications;
@@ -29,8 +30,4 @@ export default class User extends Model {
     return `${this.firstName} ${ this.lastName}`;
   }
 
-  @computed('scorecards.@each.area')
-  get areasCode() {
-    return this.scorecards.mapBy('area.code').uniq();
-  }
 }
